inject a host in the request to act like a proxy
